% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSDproject.R
\name{check_if_MSDproject}
\alias{check_if_MSDproject}
\title{Check if directory is a MSDproject}
\usage{
check_if_MSDproject(directory = getwd(), return_logical = FALSE)
}
\arguments{
\item{directory}{character. working directory (default = current working directory)}

\item{return_logical}{logical. If true will return FALSE instead of error}
}
\description{
Check if directory is a MSDproject
}
