% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{make_project}
\alias{make_project}
\title{Create new_project}
\usage{
make_project(proj_name, keep_readme_file = TRUE,
  remove_user_lib = TRUE, overwrite_rprofile = TRUE)
}
\arguments{
\item{proj_name}{character string of full path to new_project}

\item{keep_readme_file}{logical (default=TRUE) if you do not want to see readme files in project use FALSE}

\item{remove_user_lib}{logical (default=TRUE) if TRUE will attempt to remove 
user R package library from .libPaths()}

\item{overwrite_rprofile}{logical. should project .Rprofile be overwritten (default=FALSE)}
}
\description{
Creates directory structure.  User install MSDproject again in
}
