% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSDproject.R
\name{snapshot2}
\alias{snapshot2}
\title{Git commit of ctl files, SourceData and Scripts}
\usage{
snapshot2(message = "created automatic snapshot", session = TRUE, ...)
}
\arguments{
\item{message}{character. Description to be added to commit}

\item{session}{logical. Should sessionInfo be included in commit message}

\item{...}{additional arguments for git2r::commit}
}
\description{
Git commit of ctl files, SourceData and Scripts
}
